# api/openapispec/swagger.yaml
definitions:
  cluster.ClusterPayload:
    properties:
      description:
        description: ClusterDescription is the description of cluster to be created
        type: string
      displayName:
        description: ClusterDisplayName is the display name of cluster to be created
        type: string
      kubeconfig:
        description: ClusterKubeConfig is the kubeconfig of cluster to be created
        type: string
    type: object
  cluster.UploadData:
    properties:
      content:
        type: string
      fileName:
        type: string
      fileSize:
        type: integer
      sanitizedClusterContent:
        type: string
    type: object
  cluster.ValidatePayload:
    properties:
      kubeConfig:
        type: string
    type: object
...
  unstructured.Unstructured:
    properties:
      object:
        additionalProperties: true
        description: |-
          Object is a JSON compatible map with string, float, int, bool, []interface{}, or
          map[string]interface{}
          children.
        type: object
    type: object
info:
  contact: {}
  title: Karpor
  version: "1.0"
paths:
...
  /rest-api/v1/cluster/{clusterName}:
    delete:
      description: This endpoint deletes the cluster resource by name.
      parameters:
      - description: The name of the cluster
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Operation status
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete removes a cluster resource by name.
      tags:
      - cluster
    get:
      description: This endpoint returns a cluster resource by name.
      parameters:
      - description: The name of the cluster
        in: path
        name: clusterName
        required: true
        type: string
      - description: The format of the response. Either in json or yaml
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unstructured object
          schema:
            $ref: '#/definitions/unstructured.Unstructured'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get returns a cluster resource by name.
      tags:
      - cluster
    post:
      consumes:
      - text/plain
      - application/json
      description: This endpoint creates a new cluster resource using the payload.
      parameters:
      - description: cluster to create (either plain text or JSON format)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cluster.ClusterPayload'
      - description: The name of the cluster
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unstructured object
          schema:
            $ref: '#/definitions/unstructured.Unstructured'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create creates a cluster resource.
      tags:
      - cluster
    put:
      consumes:
      - text/plain
      - application/json
      description: This endpoint updates the display name and description of an existing
        cluster resource.
      parameters:
      - description: cluster to update (either plain text or JSON format)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cluster.ClusterPayload'
      - description: The name of the cluster
        in: path
        name: clusterName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unstructured object
          schema:
            $ref: '#/definitions/unstructured.Unstructured'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update updates the cluster metadata by name.
      tags:
      - cluster
  /rest-api/v1/cluster/config/file:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a KubeConfig file for cluster, with a maximum size of 2MB.
      parameters:
      - description: Upload file with field name 'file'
        in: formData
        name: file
        required: true
        type: file
      - description: cluster name
        in: formData
        name: name
        required: true
        type: string
      - description: cluster display name
        in: formData
        name: displayName
        required: true
        type: string
      - description: cluster description
        in: formData
        name: description
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Returns the content of the uploaded KubeConfig file.
          schema:
            $ref: '#/definitions/cluster.UploadData'
        "400":
          description: The uploaded file is too large or the request is invalid.
          schema:
            type: string
        "500":
          description: Internal server error.
          schema:
            type: string
      summary: Upload kubeConfig file for cluster
      tags:
      - cluster
  /rest-api/v1/cluster/config/validate:
    post:
      consumes:
      - text/plain
      - application/json
      description: Validates the provided KubeConfig using cluster manager methods.
      parameters:
      - description: KubeConfig payload to validate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/cluster.ValidatePayload'
      produces:
      - application/json
      responses:
        "200":
          description: Verification passed server version
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Validate KubeConfig
      tags:
      - cluster
  /rest-api/v1/clusters:
    get:
      description: This endpoint lists all cluster resources.
      parameters:
      - description: Whether to display summary or not. Default to false
        in: query
        name: summary
        type: boolean
      - description: The order to list the cluster. Default to order by name
        in: query
        name: orderBy
        type: string
      - description: Whether to sort the list in descending order. Default to false
        in: query
        name: descending
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of cluster objects
          schema:
            items:
              $ref: '#/definitions/unstructured.Unstructured'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: List lists all cluster resources.
      tags:
      - cluster
...
swagger: "2.0"